feat(shared-scripts): extract prepareBundledBun to reusable package
- Add packages/shared-scripts/package.json
- Extract prepare-bundled-bun.js from scripts/ to shared-scripts
- Add function signature with options: projectRoot, platform, arch, version
- Update scripts/prepareBundledBun.js as CLI wrapper calling shared module
- Enables reuse in web-cli and desktop packaging workflows
